Fix a bug where buttons behind the tutorials still display their mouse hover effects.
This commit removes the HallwayManager::doorsClickable boolean in favor of using an invisible full screen Image component on the widgetRealm and gizmoRealm that consumes all mouse interaction. Doing it this way means that the things behind the tutorial, such as the doors and jars (in the minigame) not only do not receive click events but also do not receive hover or enter/exit events. You will also not need to worry about interaction with any future things added to the scene, so long as they are behind the tutorial realms.