Add indicators above the doors to indicate what's behind them
Each indicator is implemented as a prefab that is spawned by the door at the correct time (see Door::indicatorPrefab). I didn't want these indicators to show during the tutorial so I added Hallway::StartGameplay() which is called after the tutorial (or directly after the fade, if there is no tutorial). This function finds all of the Door components using FindObjectsOfType<Door>(), and calls OnStartGameplay on them. I figure there may be other logic that we will want to run only after the tutorial that this could be useful for.