Hitting the F1 key while playing in the editor will let you instantly win the current minigame.
The #if and #endif is a preprocessor directive. These allow parts of the source code to only be compiled if a certain compile time variables are defined. In this case, UNITY_EDITOR is only true (by Unity) if the code is being compiled for the editor, and not for the html5 build.